Refactor NewFileDropdown
